income-expense-tracker/requirements.txt
streamlit
stremalit-option-menu
python-dotenv
plotly
pandas
numpy
deta
